requirements.txt
tensorflow>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
nltk>=3.8.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
wordcloud>=1.9.0
plotly>=5.15.0
streamlit>=1.25.0
jinja2>=3.1.0
aiofiles>=23.0.0
pytest>=7.4.0
black>=23.0.0
flask>=2.3.0
dash>=2.12.0
dash-bootstrap-components>=1.4.0
joblib>=1.3.0
tqdm>=4.65.0
optuna==3.5.0
